Display restore progress dialog
Addresses NuGet/Home#4225.

This change utilizes pre-existing `WaitDialogProgress` in
order to display on-build restore progress dialog when an auto-restore is
in progress.

This is to improve user experience when building hybrid solutions
containing both .NET Core and non-.NET Core projects.
